requirements.txt
Flask
Flask-CORS
requests
Frozen-Flask
python-dotenv